requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.3
six==1.15.0
tzlocal==2.1
Werkzeug==1.0.1
